Refactor 022 into same file.
